[BROKEN] Add: Entity passing to handshake
