+++ src/lib-tempo/include/tempo/entity/EntityCreationServer.hpp
@@ -1,0 +1,22 @@
+#ifndef TEMPO_ENTITY_ENTITYCREATIONSERVER_HPP
+#define TEMPO_ENTITY_ENTITYCREATIONSERVER_HPP
+
+#include <anax/World.hpp>
+
+#include <tempo/entity/EntityCreation.hpp>
+#include <tempo/entity/Render.hpp>
+#include <tempo/entity/ID.hpp>
+#include <tempo/entity/LevelManager.hpp>
+#include <tempo/entity/GridAi.hpp>
+#include <tempo/entity/PlayerLocal.hpp>
+
+namespace tempo
+{
+
+EntityCreationData dumpEntity(anax::Entity e);
+
+}
+
+#endif
+
+
