+++ src/lib-tempo/src/network/client.cpp
@@ -1,5 +1,7 @@
 #include <tempo/network/client.hpp>
 
+#include <tempo/entity/EntityCreationClient.hpp>
+
 #include <iostream>
 #include <thread>
 
@@ -94,7 +96,9 @@ void listenForServerUpdates()
 	return;
 }
 
-uint32_t handshakeHello()
+uint32_t handshakeHello(anax::World& world, 
+		        Ogre::SceneManager *scene,
+                        tempo::SystemLevelManager system_gm)
 {
 	// Package up payload
 	sf::Packet packet;
@@ -112,19 +116,22 @@ uint32_t handshakeHello()
 	// Extract Data
 	uint32_t msg = static_cast<uint32_t>(HandshakeID::DEFAULT);
 	uint32_t id = NO_CLIENT_ID;
+	uint32_t entityCount = 0;
 	packet >> msg;
 	if (msg == static_cast<uint32_t>(HandshakeID::HELLO_ROG)) {
 		packet >> id;
 		packet >> port_si;
 		packet >> port_st;
-		//
-		// TODO Extract entire level data and initialise level
-		//
+		packet >> entityCount;
+		EntityCreationData e;
+		for (int i = 0; i < entityCount; i++) {
+			packet >> e;
+			newEntity(e, world, scene, system_gm);
+		}
 	} else {
 		std::cout << "The server was rude to us when we said hello. >:("
 		          << std::endl;
 	}
-
 
 	return id;
 }
@@ -162,7 +169,11 @@ bool handshakeRoleReq(uint32_t id, ClientRole roleID, ClientRoleData &roleData)
 	return true;
 }
 	
-bool connectToAndSyncWithServer(ClientRole roleID, ClientRoleData &roleData)
+bool connectToAndSyncWithServer(ClientRole roleID, 
+                                ClientRoleData &roleData,
+                                anax::World& world,
+                                Ogre::SceneManager *scene,
+                                tempo::SystemLevelManager system_gm)
 {
 	// Bind outgoing port if not bound
 	if (sock_o.getLocalPort() == 0) {
@@ -182,7 +193,7 @@ bool connectToAndSyncWithServer(ClientRole roleID, ClientRoleData &roleData)
 		return false;
 	}
 
-	uint32_t id = handshakeHello();
+	uint32_t id = handshakeHello(world, scene, system_gm);
 	if (id == NO_CLIENT_ID) {
 		std::cout << "The server didn't like us saying HELLO!" 
 		          << std::endl;
